server/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.12
django-environ==0.11.2
celery==5.4.0
redis==5.0.4

